+++ pkg/models/order.go
@@ -14,9 +14,9 @@ type Order struct {
 	Status OrderStatus `gorm:"type:tinyint(1)" json:"status"`
 	//成交方向，以发起方（平台商用户）为准。0表示平台商用户买入，1表示平台商用户卖出。
 	Direction         int `gorm:"type:tinyint(1)" json:"direction"`
-	DistributorId     int `gorm:"type:int(11)" json:"distributor_id"`
-	MerchantId        int `gorm:"type:int(11)" json:"merchant_id"`
-	MerchantPaymentId int `gorm:"type:int(11)" json:"merchant_payment_id"`
+	DistributorId     int64 `gorm:"type:int(11)" json:"distributor_id"`
+	MerchantId        int64 `gorm:"type:int(11)" json:"merchant_id"`
+	MerchantPaymentId int64 `gorm:"type:int(11)" json:"merchant_payment_id"`
 	//扣除用户佣金金额
 	TraderCommissionAmount string `gorm:"type:varchar(32)" json:"trader_commission_amount"`
 	//扣除用户佣金币的量
